pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-infer-minimal"
version = "0.1.0"
description = "Minimal open-source inference extracted from infer.py + infer_network.py"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.5.1",
  "torchvision>=0.20.1",
  "accelerate==1.6.0",
  "diffusers==0.32.1",
  "transformers==4.54.1",
  "huggingface-hub==0.34.3",
  "safetensors==0.4.5",
  "numpy>=1.24,<2",
  "Pillow>=10.2.0",
  "tqdm==4.67.1",
  "packaging>=23.2",
  "sentencepiece==0.2.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
open-infer-json = "open_infer.run_infer_json:main"
